Flytt hentPersonerForAlleUtgjørendeVilkår og hjelpefunksjonene til UtgjørendeVilkårUtils
