+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/vedtak/VilkårUtilsTest.kt
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import java.time.LocalDate
 
-class VedtakUtilsTest {
+class VilkårUtilsTest {
 
     /**
      * Korrekt rekkefølge:
